refactor(server): extract createDirServer — agnostic fetch-native file responder
The preview file route's serving half (range, content-type, lexical guard) is
now an agnostic primitive in serveDir.ts: createDirServer(root) returns a
fetch-native (root, relPath, request) -> Response. It knows nothing about
terminals, git, or kolu — publishable as @kolu/serve-dir once a second consumer
appears. iframePreviewRoute.ts keeps only the kolu URL contract
(buildIframePreviewUrl + route-shape consts); index.ts's handler collapses to
'createDirServer(root).fetch(tail, req)'.

A 20-agent prior-art survey (PR #1219 / electricity.mdx) established no
static-serve library fits — all pipe to a Node res (breaking the Fetch-Response
+ artifact-sdk decorator composition) and none takes a per-request absolute root
or does a realpath guard — so the responder is extracted in-app, not vendored.

Per explicit direction, the realpath symlink-escape guard (kolu-git
assertRealpathUnder) is dropped for now to keep the primitive agnostic; lexical
traversal rejection (decode-then-split + path.relative containment) remains. The
guard returns later as an INJECTED guard. The dropped behavior is codified by a
serveDir.test.ts case (symlink now followed) so reinstatement has a red test to
flip.

Root stays the git repoRoot (behavior-preserving: tree, git-status decoration,
diff are repo-relative); switching the injected root to terminal $PWD
(meta.cwd) is a one-line change here, deferred as a separate product decision.

Covered by serveDir.test.ts (49 cases) + the existing code-tab preview e2e.
