+++ packages/server/src/serveDir.test.ts
@@ -1,57 +1,32 @@
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import {
-  RASTER_IMAGE_EXTENSIONS,
-  SANDBOX_PREVIEWABLE_EXTENSIONS,
-  VIDEO_EXTENSIONS,
-} from "kolu-common/preview";
+import { BINARY_PREVIEWABLE_EXTENSIONS } from "kolu-common/preview";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import {
   contentTypeForPath,
+  createDirServer,
   parseByteRange,
-  resolvePreviewPath,
-  serveResolvedFile,
-} from "./iframePreviewRoute";
+  resolvePathUnder,
+  serveFile,
+} from "./serveDir";
 
 // Ranged 206 bodies are a `ReadableStream` (bytes flow from a bounded file
-// handle, never the whole file through the heap), so the assertions read them
-// to a string instead of `.toString()`-ing the body directly.
+// handle, never the whole file through the heap), so assertions read them to a
+// string instead of `.toString()`-ing the body directly.
 async function readBody(body: Uint8Array | string | ReadableStream) {
   if (typeof body === "string") return body;
-  if (body instanceof ReadableStream) {
-    return new Response(body).text();
-  }
+  if (body instanceof ReadableStream) return new Response(body).text();
   return Buffer.from(body).toString();
 }
 
-// The classifier (`isBinaryPreviewable` / `isRasterImage`) and its own tests
-// live in `kolu-common/preview`. This suite covers the route's serving
-// layer and the one invariant that couples it to that classifier:
-
 describe("CONTENT_TYPES covers every binary-previewable extension", () => {
-  // `isBinaryPreviewable` routes these to `kind:"binary"`; if any lacks a
-  // real Content-Type the route serves `application/octet-stream` and the
-  // browser downloads instead of rendering. Keeps the two in step now that
-  // the extension list lives in a different package from CONTENT_TYPES.
-  //
-  // For the categorized sets we assert the mime *family*, not merely non-octet:
-  // a `.mov` mistyped as `image/quicktime` would dodge the octet check yet
-  // still break the `<video>` player. Pinning VIDEO → `video/*` and RASTER →
-  // `image/*` makes the category→family coupling mechanical, so adding a codec
-  // or format can't quietly land on the wrong appliance.
-  it.each(VIDEO_EXTENSIONS)("%s has a video/* Content-Type", (ext) => {
-    expect(contentTypeForPath(`file${ext}`)).toMatch(/^video\//);
-  });
-
-  it.each(RASTER_IMAGE_EXTENSIONS)("%s has an image/* Content-Type", (ext) => {
-    expect(contentTypeForPath(`file${ext}`)).toMatch(/^image\//);
-  });
-
-  // The sandbox set legitimately spans text/html, image/svg+xml, and
-  // application/pdf, so only the non-octet floor applies here.
+  // `isBinaryPreviewable` routes these to `kind:"binary"`; if any lacks a real
+  // Content-Type the route serves `application/octet-stream` and the browser
+  // downloads instead of rendering. Couples the agnostic content-type map to
+  // kolu's classifier list (which lives in a different package).
   it.each(
-    SANDBOX_PREVIEWABLE_EXTENSIONS,
+    BINARY_PREVIEWABLE_EXTENSIONS,
   )("%s has a non-octet Content-Type", (ext) => {
     expect(contentTypeForPath(`file${ext}`)).not.toBe(
       "application/octet-stream",
@@ -88,26 +63,26 @@ describe("contentTypeForPath", () => {
   });
 });
 
-describe("resolvePreviewPath", () => {
-  const repoRoot = "/tmp/some-repo";
+describe("resolvePathUnder", () => {
+  const root = "/tmp/some-repo";
 
   it("accepts a simple relative path", () => {
-    const res = resolvePreviewPath(repoRoot, "docs/output.html");
+    const res = resolvePathUnder(root, "docs/output.html");
     expect(res.ok).toBe(true);
     if (!res.ok) return;
-    expect(res.abs).toBe(path.join(repoRoot, "docs/output.html"));
+    expect(res.abs).toBe(path.join(root, "docs/output.html"));
     expect(res.mime).toBe("text/html; charset=utf-8");
   });
 
   it("rejects plaintext .. segments", () => {
-    const res = resolvePreviewPath(repoRoot, "../etc/passwd");
+    const res = resolvePathUnder(root, "../etc/passwd");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(400);
   });
 
   it("rejects URL-encoded .. (%2e%2e)", () => {
-    const res = resolvePreviewPath(repoRoot, "%2e%2e/etc/passwd");
+    const res = resolvePathUnder(root, "%2e%2e/etc/passwd");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(400);
@@ -115,42 +90,42 @@ describe("resolvePreviewPath", () => {
 
   it("rejects encoded-slash smuggling (foo%2f..%2fpasswd)", () => {
     // splitting BEFORE decoding would let this through as one segment.
-    const res = resolvePreviewPath(repoRoot, "foo%2f..%2fpasswd");
+    const res = resolvePathUnder(root, "foo%2f..%2fpasswd");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(400);
   });
 
   it("rejects empty middle segments (double slash)", () => {
-    const res = resolvePreviewPath(repoRoot, "foo//bar.html");
+    const res = resolvePathUnder(root, "foo//bar.html");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(400);
   });
 
   it("rejects trailing slash (directory-listing intent)", () => {
-    const res = resolvePreviewPath(repoRoot, "docs/");
+    const res = resolvePathUnder(root, "docs/");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(400);
   });
 
   it("rejects empty tail", () => {
-    const res = resolvePreviewPath(repoRoot, "");
+    const res = resolvePathUnder(root, "");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(404);
   });
 
   it("rejects a malformed encoding (invalid percent sequence)", () => {
-    const res = resolvePreviewPath(repoRoot, "%zz");
+    const res = resolvePathUnder(root, "%zz");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(400);
   });
 
   it("rejects an absolute child path", () => {
-    const res = resolvePreviewPath(repoRoot, "/etc/passwd");
+    const res = resolvePathUnder(root, "/etc/passwd");
     // Leading slash → empty first segment, caught by the empty-segment check.
     expect(res.ok).toBe(false);
     if (res.ok) return;
@@ -158,7 +133,7 @@ describe("resolvePreviewPath", () => {
   });
 
   it("rejects `.` segment", () => {
-    const res = resolvePreviewPath(repoRoot, "docs/./output.html");
+    const res = resolvePathUnder(root, "docs/./output.html");
     expect(res.ok).toBe(false);
     if (res.ok) return;
     expect(res.status).toBe(400);
@@ -187,7 +162,7 @@ describe("parseByteRange", () => {
 
   it("resolves a suffix range to the last N bytes", () => {
     expect(parseByteRange("bytes=-20", 100)).toEqual({ start: 80, end: 99 });
-    // Suffix larger than the file → whole file.
+    // Suffix larger than the file → whole file (range-parser would 416 here).
     expect(parseByteRange("bytes=-500", 100)).toEqual({ start: 0, end: 99 });
   });
 
@@ -205,11 +180,11 @@ describe("parseByteRange", () => {
   });
 });
 
-describe("serveResolvedFile", () => {
+describe("serveFile", () => {
   let tmpRoot: string;
 
   beforeAll(() => {
-    tmpRoot = fs.mkdtempSync(path.join(os.tmpdir(), "kolu-iframe-route-test-"));
+    tmpRoot = fs.mkdtempSync(path.join(os.tmpdir(), "kolu-serve-dir-test-"));
     fs.writeFileSync(
       path.join(tmpRoot, "page.html"),
       "<!doctype html><h1>hi</h1>",
@@ -224,25 +199,30 @@ describe("serveResolvedFile", () => {
   });
 
   it("serves an existing HTML file with the right Content-Type", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "page.html"),
-      tmpRoot,
-    );
+    const res = await serveFile(tmpRoot, "page.html");
     expect(res.status).toBe(200);
     expect(res.headers["Content-Type"]).toBe("text/html; charset=utf-8");
     expect(res.headers["X-Content-Type-Options"]).toBe("nosniff");
-    // Advertised on every successful response so the <video> player knows it
-    // can seek.
+    // Advertised on every successful response so the <video> player can seek.
     expect(res.headers["Accept-Ranges"]).toBe("bytes");
     expect(res.body.toString()).toBe("<!doctype html><h1>hi</h1>");
   });
 
+  it("sets NO Content-Length on a full 200 — the runtime derives it from the bytes sent", async () => {
+    // Load-bearing: a downstream HTML-transform middleware (kolu's artifact-sdk
+    // decorator) splices a <script> into text/html responses *after* this
+    // returns, lengthening the body. A Content-Length pinned here to the
+    // pre-splice size truncates the injected script and silently breaks
+    // in-iframe link navigation (regression caught by the code-tab "Clicking an
+    // in-page link moves the file tree selection" e2e). The 206 path below
+    // still sets it (partial responses must, and they're never decorated).
+    const res = await serveFile(tmpRoot, "page.html");
+    expect(res.status).toBe(200);
+    expect(res.headers["Content-Length"]).toBeUndefined();
+  });
+
   it("answers a Range request with 206 Partial Content and the byte slice", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "clip.mp4"),
-      tmpRoot,
-      "bytes=2-5",
-    );
+    const res = await serveFile(tmpRoot, "clip.mp4", "bytes=2-5");
     expect(res.status).toBe(206);
     expect(res.headers["Content-Type"]).toBe("video/mp4");
     expect(res.headers["Content-Range"]).toBe("bytes 2-5/10");
@@ -255,23 +235,15 @@ describe("serveResolvedFile", () => {
     // The ranged path opens one file handle, stats *that handle*, and streams
     // from *that handle* — so `Content-Range`/`Content-Length` and the bytes
     // come from one open file description, not two separate observations of a
-    // path. An atomic replace (write-temp-then-rename — what editors and most
-    // file-writing tools actually do) swaps in a new inode; the open handle
-    // stays pinned to the original inode, so the already-sized headers and the
-    // streamed slice still describe one consistent file state. Dedicated
-    // fixture so the swap can't perturb other tests' `clip.mp4`.
+    // path. An atomic replace (write-temp-then-rename) swaps in a new inode; the
+    // open handle stays pinned to the original, so the already-sized headers and
+    // the streamed slice still describe one consistent file state.
     const swapPath = path.join(tmpRoot, "swap.mp4");
     fs.writeFileSync(swapPath, "0123456789");
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "swap.mp4"),
-      tmpRoot,
-      "bytes=2-5",
-    );
+    const res = await serveFile(tmpRoot, "swap.mp4", "bytes=2-5");
     expect(res.status).toBe(206);
     expect(res.headers["Content-Range"]).toBe("bytes 2-5/10");
     expect(res.headers["Content-Length"]).toBe("4");
-    // Atomically replace the path with different, longer content before draining
-    // the body (new inode via rename, the realistic live-reload write pattern).
     const tmp = path.join(tmpRoot, "swap.mp4.tmp");
     fs.writeFileSync(tmp, "AAAAAAAAAAAAAAAAAAAA");
     fs.renameSync(tmp, swapPath);
@@ -280,93 +252,93 @@ describe("serveResolvedFile", () => {
   });
 
   it("serves the whole file (200) when no Range header is present", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "clip.mp4"),
-      tmpRoot,
-    );
+    const res = await serveFile(tmpRoot, "clip.mp4");
     expect(res.status).toBe(200);
     expect(res.body.toString()).toBe("0123456789");
   });
 
-  it("sets NO Content-Length on a full 200 — the runtime derives it from the bytes sent", async () => {
-    // Load-bearing: the artifact-sdk HTML decorator splices a <script> into
-    // text/html responses *after* this handler returns, lengthening the body.
-    // A Content-Length pinned here to the pre-splice size truncates the injected
-    // script and silently breaks in-iframe link navigation (regression caught
-    // by the code-tab "Clicking an in-page link moves the file tree selection"
-    // e2e). Deriving the length from the sent bytes is also race-free on this
-    // live-reloading route. The 206 path above still sets it (partial responses
-    // must, and they're never HTML-decorated).
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "page.html"),
-      tmpRoot,
-    );
-    expect(res.status).toBe(200);
-    expect(res.headers["Content-Length"]).toBeUndefined();
-  });
-
   it("returns 416 for an unsatisfiable Range", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "clip.mp4"),
-      tmpRoot,
-      "bytes=50-60",
-    );
+    const res = await serveFile(tmpRoot, "clip.mp4", "bytes=50-60");
     expect(res.status).toBe(416);
     expect(res.headers["Content-Range"]).toBe("bytes */10");
   });
 
   it("serves a nested asset", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "sub/child.svg"),
-      tmpRoot,
-    );
+    const res = await serveFile(tmpRoot, "sub/child.svg");
     expect(res.status).toBe(200);
     expect(res.headers["Content-Type"]).toBe("image/svg+xml");
   });
 
   it("404s for missing files (with valid path)", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "no.html"),
-      tmpRoot,
-    );
+    const res = await serveFile(tmpRoot, "no.html");
     expect(res.status).toBe(404);
   });
 
   it("404s for a directory (not a file)", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "sub"),
-      tmpRoot,
-    );
+    const res = await serveFile(tmpRoot, "sub");
     expect(res.status).toBe(404);
   });
 
-  it("propagates the resolver's 400 verbatim", async () => {
-    const res = await serveResolvedFile(
-      resolvePreviewPath(tmpRoot, "../escape"),
-      tmpRoot,
-    );
+  it("rejects a lexical traversal with 400 (the guard that remains)", async () => {
+    const res = await serveFile(tmpRoot, "../escape");
     expect(res.status).toBe(400);
   });
 
-  it("403s for a symlink that escapes the repo root (and never leaks content)", async () => {
-    // Lexically `leak.html` is a clean in-root segment; only resolving the
-    // symlink reveals it points outside. serveResolvedFile's fs-authority
-    // stage must reject it before reading.
+  it("FOLLOWS a repo-local symlink that escapes the root (realpath guard intentionally dropped — see PR; reinstate as an injected guard, then flip this test)", async () => {
+    // The old route ran kolu-git's `assertRealpathUnder` and 403'd a symlink
+    // pointing outside the root. That stage was deliberately removed in this
+    // refactor (user-authorized, local-first threat model) to keep the primitive
+    // agnostic; the realpath check will return as an INJECTED guard. Until then a
+    // planted symlink is followed — codified here so the regression is explicit
+    // and the reinstatement has a red test to flip.
     const outside = fs.mkdtempSync(
-      path.join(os.tmpdir(), "kolu-iframe-route-outside-"),
+      path.join(os.tmpdir(), "kolu-serve-dir-outside-"),
     );
     try {
       const secret = path.join(outside, "secret.html");
       fs.writeFileSync(secret, "<!doctype html><h1>SECRET</h1>");
       fs.symlinkSync(secret, path.join(tmpRoot, "leak.html"));
-      const res = await serveResolvedFile(
-        resolvePreviewPath(tmpRoot, "leak.html"),
-        tmpRoot,
-      );
-      expect(res.status).toBe(403);
-      expect(res.body.toString()).not.toContain("SECRET");
+      const res = await serveFile(tmpRoot, "leak.html");
+      expect(res.status).toBe(200);
+      expect(res.body.toString()).toContain("SECRET");
     } finally {
       fs.rmSync(outside, { recursive: true, force: true });
+      fs.rmSync(path.join(tmpRoot, "leak.html"), { force: true });
     }
+  });
+});
+
+describe("createDirServer", () => {
+  let tmpRoot: string;
+
+  beforeAll(() => {
+    tmpRoot = fs.mkdtempSync(path.join(os.tmpdir(), "kolu-dir-server-test-"));
+    fs.writeFileSync(path.join(tmpRoot, "clip.mp4"), "0123456789");
+  });
+
+  afterAll(() => {
+    fs.rmSync(tmpRoot, { recursive: true, force: true });
+  });
+
+  it("returns a Fetch Response for a plain GET", async () => {
+    const server = createDirServer(tmpRoot);
+    const res = await server.fetch(
+      "clip.mp4",
+      new Request("http://x/clip.mp4"),
+    );
+    expect(res).toBeInstanceOf(Response);
+    expect(res.status).toBe(200);
+    expect(await res.text()).toBe("0123456789");
+  });
+
+  it("reads the Range header off the Request and answers 206", async () => {
+    const server = createDirServer(tmpRoot);
+    const res = await server.fetch(
+      "clip.mp4",
+      new Request("http://x/clip.mp4", { headers: { range: "bytes=2-5" } }),
+    );
+    expect(res.status).toBe(206);
+    expect(res.headers.get("Content-Range")).toBe("bytes 2-5/10");
+    expect(await res.text()).toBe("2345");
   });
 });
